requirements.txt
# ============================================================================
# MUSE: Resolving Manifold Misalignment in Visual Tokenization
via Topological Orthogonality
# Core Dependencies
# ============================================================================

# PyTorch (install separately for your CUDA version: https://pytorch.org)
# torch>=2.1.0
# torchvision>=0.16.0

# Training & Distributed
accelerate>=1.0.0
omegaconf>=2.1.0

# Transformers & Models
transformers>=4.37.0
diffusers>=0.33.0
timm>=1.0.0
open-clip-torch>=2.31.0

# Data Loading
webdataset>=0.2.100
braceexpand>=0.1.7
datasets>=3.1.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.1.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
einops>=0.8.0

# Evaluation
torch-fidelity>=0.3.0
torchinfo>=1.8.0

# Logging & Utilities
tqdm>=4.60.0
termcolor>=2.0.0
wandb>=0.15.0

# Misc
requests>=2.30.0
gdown>=5.0.0
